+++ src/main/java/me/lucaspmntl/geofreight/dto/melhorenvio/response/MelhorEnvioResponseDTO.java
@@ -3,13 +3,15 @@ package me.lucaspmntl.geofreight.dto.melhorenvio.response;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record MelhorEnvioResponseDTO(
         @JsonProperty("name")
         String transportName,
 
         @JsonProperty("price")
-        Double transportCompanyPrice,
+        BigDecimal transportCompanyPrice,
 
         @JsonProperty("delivery_time")
         Integer deliveryTime,
